app/requirements.txt
Flask
mysql-connector
SQLAlchemy==1.3.5
PyMySQL==0.9.3

#appdirs==1.4.3
#astroid==2.2.5
#attrs==19.1.0
#Automat==0.7.0
#autopep8==1.4.4
#blinker==1.4
#Blueprints==2.3.0.2
#Click==7.0
#constantly==15.1.0
#coverage==4.5.4
#Flask==1.0.3
#Flask-Mail==0.9.1
#Flask-Script==2.0.6
#flask-sqlacodegen==1.1.6.1
#Flask-SQLAlchemy==2.4.0
#httpagentparser==1.9.0
#incremental==17.5.0
#inflect==2.1.0
#isort==4.3.21
#itsdangerous==1.1.0
#Jinja2==2.10.1
#lazy-object-proxy==1.4.1
#Mako==1.0.13
#MarkupSafe==1.1.1
#mccabe==0.6.1
#pep8==1.7.1
#piwikapi==0.3
#pycodestyle==2.5.0
#pycrypto==2.6.1
#pylint==2.3.1
#six==1.12.0
#sqlparse==0.3.0
#sqltap==0.3.11
#typed-ast==1.4.0
#ujson==1.35
#Werkzeug==0.15.4
#wrapt==1.11.2
#zope.interface==4.6.0
#